test(ai): move ChatGPT SSE EOF tail cases under the max-lines budget
